sizes are padded correctly, outputs are still bad
